Utilities for creating diagrams using Tikz
This patch adds various LaTeX styles and commands for creating diagrams.
Each diagram will be placed in a dedicated 'img/<diagram>.tikz' file. Such
a file contains solely the actual tikz image. It can either be included
by the manual or processed into a PDF file. For the latter, the
'img/Makefile' can be used.

As a first example, the patch comes with the diagram 'object_identity',
which illustrates the relationship between an RPC object, its
capability, the capability space of the protection domain, and a
kernel-protected object identity.